Add additional options and reorganize UI code
