--- src/popup/options/options.css
@@ -1,72 +1,0 @@
-html {
-    overflow: hidden;
-}
-
-.panel {
-    background-color: var(--mullvad-blue);
-    height: 100%;
-    left: 0;
-    position: absolute;
-    top: 100%;
-    width: 100%;
-    overflow-y: auto;
-    z-index: 9999;
-}
-
-.panel.panel--visible {
-    top: 0;
-}
-
-.panel__header {
-    background-color: var(--mullvad-dark-blue);
-    align-items: center;
-    display: flex;
-    gap: 10px;
-    padding: 10px 20px;
-}
-
-.panel__title {
-    font-size: 1.2em;
-    flex: 1;
-    margin: initial;
-}
-
-.options {
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-    padding: 20px;
-}
-
-.options > hr {
-    border: initial;
-    border-top: 1px solid rgba(255, 255, 255, 0.25);
-    margin: initial;
-    width: 100%;
-}
-
-.option {
-    display: flex;
-    flex-direction: column;
-    gap: 5px;
-}
-.option.option--inline {
-    align-items: center;
-    flex-wrap: wrap;
-    flex-direction: row;
-}
-
-.option__control {
-    display: flex;
-}
-.option__control > input[type="text"] {
-    width: 100%;
-}
-
-
-.option__description {
-    width: 100%;
-    font-size: 0.9em;
-    opacity: 0.75;
-}
-
