+++ src/ui/options/index.html
@@ -1,0 +1,15 @@
+<!doctype html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+
+        <link rel="stylesheet" href="../shared.css">
+        <link rel="stylesheet" href="options.css">
+
+        <script src="../../browser-polyfill.min.js" defer></script>
+        <script src="options.bundle.js" defer></script>
+    </head>
+    <body>
+        <!-- App content -->
+    </body>
+</html>
