+++ src/ui/options/options-firefox.css
@@ -1,0 +1,15 @@
+body {
+    background-color: white;
+}
+
+@media (prefers-color-scheme:dark) {
+    body {
+        background-color: #202023;
+    }
+}
+
+@media (prefers-color-scheme: light) {
+    input[type="checkbox"]:checked::before {
+        filter: invert(1);
+    }
+}
