+++ src/ui/popup/index.html
@@ -3,11 +3,11 @@
     <head>
         <meta charset="UTF-8">
 
+        <link rel="stylesheet" href="../shared.css">
+        <link rel="stylesheet" href="../options/options.css">
         <link rel="stylesheet" href="style.css">
-        <link rel="stylesheet" href="widgets.css">
-        <link rel="stylesheet" href="options/options.css">
 
-        <script src="../browser-polyfill.min.js" defer></script>
+        <script src="../../browser-polyfill.min.js" defer></script>
         <script src="popup.bundle.js" defer></script>
     </head>
     <body>
