+++ src/ui/popup/options/options.css
@@ -1,0 +1,49 @@
+:root {
+    --border-color: rgba(255, 255, 255, 0.25);
+}
+
+.options {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    padding: 20px;
+}
+
+.options > hr {
+    border: initial;
+    border-top: 1px solid var(--border-color);
+    margin: initial;
+    width: 100%;
+}
+
+.option {
+    display: flex;
+    flex-direction: column;
+    gap: 5px;
+    pointer-events: none;
+}
+.option.option--inline {
+    align-items: center;
+    flex-wrap: wrap;
+    flex-direction: row;
+}
+
+.option__control,
+.option__label {
+    pointer-events: initial;
+}
+
+.option__control {
+    display: flex;
+}
+.option__control > input[type="text"] {
+    width: 100%;
+}
+
+
+.option__description {
+    width: 100%;
+    font-size: 0.85em;
+    opacity: 0.75;
+}
+
