+++ src/ui/popup/style-mac.css
